# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Activity Log Controller API
  description: |
    API RESTful para o sistema de controle de atividades e log de funcionários.
    Esta API permite gerenciar autenticação de usuários, cadastro de funcionários,
    registro de atividades e envio de feedbacks entre supervisores e fiscais.
  version: 1.0.0
  contact:
    name: Equipe de Desenvolvimento
    email: suporte@activitylogcontroller.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Servidor de desenvolvimento local
  - url: https://activity-log-controller.onrender.com/api
    description: Servidor de produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
        message:
          type: string
          description: Detalhes adicionais do erro
        fields:
          type: object
          description: Campos que falharam na validação
    User:
      type: object
      properties:
        uid:
          type: string
          description: ID único do usuário no Firebase Auth
        email:
          type: string
          format: email
          description: Email do usuário
        name:
          type: string
          description: Nome completo do usuário
        picture:
          type: string
          format: uri
          nullable: true
          description: URL da foto de perfil
        provider:
          type: string
          enum: [email, google]
          description: Provedor de autenticação
        role:
          type: string
          enum: [fiscal, supervisor]
          description: Papel do usuário no sistema
        emailVerified:
          type: boolean
          description: Indica se o email foi verificado
...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login realizado com sucesso
        token:
          type: string
          description: Token JWT da API para autenticação em endpoints protegidos
        sessionId:
          type: string
          description: ID da sessão criada
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /auth/google:
    post:
      tags:
        - Autenticação
      summary: Login com Google
      description: |
        Autentica um usuário usando o idToken do Google obtido através do Firebase Client SDK.
        Cria ou atualiza o perfil do usuário no Firestore.
      operationId: loginWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idToken
              properties:
                idToken:
                  type: string
                  description: Token ID do Google obtido via Firebase Client SDK
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/me:
    get:
      tags:
        - Autenticação
      summary: Obter usuário autenticado
      description: |
        Retorna os dados do usuário atualmente autenticado.
        Requer autenticação via Bearer token.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...